split handlers
