+++ src/handlers/fromBody.ts
@@ -1,0 +1,10 @@
+import { IRouterContext } from 'koa-router'
+
+
+export const fromBody = (ctx: IRouterContext, fieldName: string, type: 'string' | 'number' | 'boolean') => {
+  const value = ctx.request.body[fieldName]
+  if (!value || typeof value !== type) {
+    throw { status: 400, message: `Body must include ${fieldName}, a ${type}` }
+  }
+  return value
+}
